Context API: Typescript with useContext
-created a file for context component
-moved addToCart and removeFromCart to context
-moved types to separate file
