+++ src/index.tsx
@@ -3,6 +3,7 @@ import ReactDOM from "react-dom/client";
 
 import App from "./App";
 import { QueryClient, QueryClientProvider } from "react-query";
+import ContextProvider from "./Context/Context";
 
 const client = new QueryClient();
 
@@ -10,7 +11,9 @@ const root = ReactDOM.createRoot(
 	document.getElementById("root") as HTMLElement
 );
 root.render(
-	<QueryClientProvider client={client}>
-		<App />
-	</QueryClientProvider>
+	<ContextProvider>
+		<QueryClientProvider client={client}>
+			<App />
+		</QueryClientProvider>
+	</ContextProvider>
 );
